githubEventsProducer/requirements.txt
aiohttp==3.9.0b0
confluent-kafka
PyGithub
python-dotenv